+++ src/definition.ts
@@ -1,4 +1,4 @@
-import { ASTBase, ASTIdentifier, ASTMemberExpression } from 'greyscript-core';
+import { ASTBase, ASTIdentifier, ASTIndexExpression, ASTMemberExpression } from 'greyscript-core';
 import vscode, {
   CancellationToken,
   DefinitionLink,
@@ -8,7 +8,7 @@ import vscode, {
   TextDocument
 } from 'vscode';
 
-import ASTStringify from './helper/ast-stringify';
+import transformASTToNamespace from './helper/ast-namespace';
 import documentParseQueue from './helper/document-manager';
 import { LookupHelper } from './helper/lookup-type';
 
@@ -67,8 +67,11 @@ export function activate(_context: ExtensionContext) {
       const previous = outer.length > 0 ? outer[outer.length - 1] : undefined;
       let identifer = closest.name;
 
-      if (previous && previous instanceof ASTMemberExpression) {
-        identifer = ASTStringify(previous);
+      if (previous && (
+        previous instanceof ASTMemberExpression ||
+        previous instanceof ASTIndexExpression
+      )) {
+        identifer = transformASTToNamespace(previous);
       }
 
       const definitions = findAllDefinitions(helper, identifer, closest.scope!);