+++ src/helper/type-manager.ts
@@ -23,7 +23,8 @@ import {
 } from 'greyscript-meta';
 import { TextDocument } from 'vscode';
 
-import ASTStringify from './ast-stringify';
+import transformASTToNamespace from './ast-namespace';
+import transformASTToString from './ast-stringify';
 
 export class TypeInfo {
   label: string;
@@ -94,15 +95,15 @@ export class TypeMap {
 
   lookupTypeOfNamespace(item: ASTBase): TypeInfo | null {
     const me = this;
-    const name = ASTStringify(item);
+    const namespace = transformASTToNamespace(item);
     let currentScope = item.scope;
 
     while (currentScope) {
       if (me.refs.has(currentScope)) {
         const typeMap = me.refs.get(currentScope)!;
 
-        if (typeMap.has(name)) {
-          const typeInfo = typeMap.get(name)!;
+        if (typeMap.has(namespace)) {
+          const typeInfo = typeMap.get(namespace)!;
           return typeInfo;
         }
       }
@@ -277,7 +278,7 @@ export class TypeMap {
       arguments: item.parameters.map((arg: ASTBase) => {
         if (arg.type === ASTType.Identifier) {
           return {
-            label: ASTStringify(arg),
+            label: transformASTToString(arg),
             type: 'any'
           } as SignatureDefinitionArg;
         }
@@ -285,7 +286,7 @@ export class TypeMap {
         const assignment = arg as ASTAssignmentStatement;
 
         return {
-          label: ASTStringify(assignment.variable),
+          label: transformASTToString(assignment.variable),
           type: me.resolve(assignment.init)?.type[0] || 'any'
         };
       }),
@@ -399,7 +400,7 @@ export class TypeMap {
     me.refs.set(scope, identiferTypes);
 
     for (const assignment of assignments) {
-      const name = ASTStringify(assignment.variable);
+      const name = transformASTToNamespace(assignment.variable);
       const resolved = me.resolve(assignment.init);
 
       if (resolved === null) continue;