+++ src/helper/lookup-type.ts
@@ -4,10 +4,7 @@ import {
   ASTBase,
   ASTBaseBlockWithScope,
   ASTChunk,
-  ASTFunctionStatement,
-  ASTIdentifier,
   ASTIndexExpression,
-  ASTMapConstructorExpression,
   ASTMemberExpression,
   ASTType
 } from 'greyscript-core';