+++ CHANGELOG.md
@@ -481,3 +481,4 @@ All notable changes to this project will be documented in this file.
 - fix tooltip for lastIndexOf
 - remove usage of boolean type in tooltips and signatures to avoid confusion since technically booleans do not exist
 - minor improvement to index expression type analyzing
+- use modified transformer to get namespace for type