+++ CHANGELOG.md
@@ -484,3 +484,5 @@ All notable changes to this project will be documented in this file.
 - use modified transformer to get namespace for type
 - partialy support type resolve from identifier assigned in imported file
 - fix type resolve within slice expression
+- improve type resolve for assignments using locals, globals or outer prefix
+- extend namespace find method to use assignment instead of namespaces coming from parser