+++ CHANGELOG.md
@@ -486,3 +486,4 @@ All notable changes to this project will be documented in this file.
 - fix type resolve within slice expression
 - improve type resolve for assignments using locals, globals or outer prefix
 - extend namespace find method to use assignment instead of namespaces coming from parser
+- fix possibly wrong start position of member, index and call expression