+++ CHANGELOG.md
@@ -489,3 +489,4 @@ All notable changes to this project will be documented in this file.
 - fix possibly wrong start position of member, index and call expression
 - inject map constructor namespaces
 - inject list constructor namespaces
+- add block comment support